Tests/LibWeb/Text/expected/css/preferred-color-scheme-auto-reset.txt

Forced dark: true
Reset to auto: false
